# Conflicts:
#	NEWS.md
#	doc/documentation.md
#	src/test/java/org/folio/entlinks/controller/LinksStatisticsInstanceIT.java
#	src/test/java/org/folio/support/base/IntegrationTestBase.java

+++ src/test/java/org/folio/support/TestDataUtils.java
@@ -283,7 +283,7 @@ public class TestDataUtils {
           .naturalId(naturalId)
           .build())
         .linkingRule(InstanceAuthorityLinkingRule.builder()
-          .id((long) TAGS_TO_RULE_IDS.get(tag))
+          .id(TAGS_TO_RULE_IDS.get(tag))
           .bibField(tag)
           .build())
         .status(status)
